refactored the card details section
refactored the card details section
