+++ src/components/cardhelper/renderDetails/renderDetails.scss
@@ -1,0 +1,34 @@
+.details {
+	display: flex;
+	gap: 1.9rem;
+	height: 1.6rem;
+	margin-top: 0.8rem;
+	margin-bottom: 0.5rem;
+	align-items: center;
+
+	> div {
+		position: relative;
+		//except last child, all the child
+		&:not(:last-child) {
+			&::after {
+				position: absolute;
+				display: inline-block;
+				content: '';
+				top: 50%;
+				transform: translateY(-50%);
+				border-radius: 50%;
+				right: -1.1rem;
+				width: 0.3rem;
+				height: 0.3rem;
+				background-color: #fff;
+			}
+		}
+	}
+	.detail-category {
+		display: flex;
+		align-items: center;
+		img {
+			margin-right: 0.6rem;
+		}
+	}
+}
